pyproject.toml
[tool.poetry]
name = "web-osi-intelligence-analyzer"
version = "0.1.0"
description = "A production-ready, secure, scalable web application for complete OSI + TCP/IP layer analysis."
authors = ["Antigravity <antigravity@deepmind.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
gunicorn = "^21.2.0"
pydantic = {extras = ["email"], version = "^2.6.0"}
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
dnspython = "^2.6.0"
psutil = "^5.9.8"
netifaces = "^0.11.0"
aiohttp = "^3.9.3"
scapy = "^2.5.0"
jinja2 = "^3.1.3"
python-multipart = "^0.0.9"
email-validator = "^2.1.0.post1"
httpx = "^0.26.0"
websockets = "^12.0"
geoip2 = "^4.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
